# flate: Faster token writer, 1-2% on fastest