```
λ benchcmp before.txt after.txt
benchmark                            old ns/op     new ns/op     delta
BenchmarkEncodeDigitsSpeed1e4-32     48523         47923         -1.24%
BenchmarkEncodeDigitsSpeed1e5-32     642126        632069        -1.57%
BenchmarkEncodeDigitsSpeed1e6-32     6197352       6046440       -2.44%
BenchmarkEncodeTwainSpeed1e4-32      70716         69414         -1.84%
BenchmarkEncodeTwainSpeed1e5-32      641631        632854        -1.37%
BenchmarkEncodeTwainSpeed1e6-32      6432128       6323973       -1.68%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkEncodeDigitsSpeed1e4-32     206.09       208.67       1.01x
BenchmarkEncodeDigitsSpeed1e5-32     155.73       158.21       1.02x
BenchmarkEncodeDigitsSpeed1e6-32     161.36       165.39       1.02x
BenchmarkEncodeTwainSpeed1e4-32      141.41       144.06       1.02x
BenchmarkEncodeTwainSpeed1e5-32      155.85       158.01       1.01x
BenchmarkEncodeTwainSpeed1e6-32      155.47       158.13       1.02x
```